added ntfy for notifications, fixed bugs and made duplication filter logic more efficient
